% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pckg.cheb.r
\name{cheb.1st}
\alias{cheb.1st}
\title{Generating Chebyshev Polynomials of first kind}
\usage{
cheb.1st(x.axis, n)
}
\description{
cheb.1st erzeugt Chebyshev Polynome erster Art aus beliebiger X-Achse
}
\examples{
cheb.1st(x.axis, n)
}

