% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pckg.cheb.r
\name{cheb.2nd}
\alias{cheb.2nd}
\title{Generating Chebyshev Polynomials of second kind}
\usage{
cheb.2nd(x.axis, n)
}
\description{
cheb.2nd erzeugt Chebyshev Polynome zweiter Art aus beliebiger X-Achse
}
\examples{
cheb.2nd(x.axis, n)
}

