% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pckg.cheb.r
\name{cheb.deriv.1st}
\alias{cheb.deriv.1st}
\title{Calculation of the values of the first derivation}
\usage{
cheb.deriv.1st(x.axis, cheb.coeff)
}
\description{
cheb.deriv.1st berechnet aus den Chebyshev-Koeffizienten die Werte der ersten Ableitung
}
\examples{
cheb.deriv.1st(x.axis, cheb.coeff)
}

