% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pckg.cheb.r
\name{cheb.deriv.2nd}
\alias{cheb.deriv.2nd}
\title{Calculation of the values of the second derivation}
\usage{
cheb.deriv.2nd()
}
\description{
cheb.deriv.2nd berechnet aus den Chebyshev-Koeffizienten die Werte der zweiten Ableitung
}
\examples{
##
}

