% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pckg.cheb.r
\name{cheb.fit}
\alias{cheb.fit}
\title{Curve Fitting with Chebyshev Polynomials}
\usage{
cheb.fit(d, x.axis, n)
}
\description{
\code{cheb.fit} fittet ein Chebyshev-Polynom beliebiger Ordnung an einen Datensatz/Zeitreihe mittels Least Squares Verfahren
}
\examples{
cheb.fit(d, x.axis, n)
}

