% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pckg.cheb.r
\name{cheb.rescale}
\alias{cheb.rescale}
\title{Rescaling of Y-Axis}
\usage{
cheb.rescale(x.cheb, x.axis)
}
\description{
cheb.rescale reskaliert die für den Fit erzeugte Achse auf die Ursprüngliche
}
\examples{
cheb.rescale(x.cheb, x.axis)
}

