% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pckg.cheb.r
\name{cheb.scale}
\alias{cheb.scale}
\title{Scaling of Y-Axis}
\usage{
cheb.scale(x.axis)
}
\description{
\code(cheb.scale) skaliert beliebige Y-Achse auf Achse, die für Polynom-fits verträglich ist.
}
\examples{
x.axis <- c(0:30)
cheb.scale(x.axis)
}

