% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pckg.cheb.r
\name{cheb.val}
\alias{cheb.val}
\title{Calculation of Values of the model fit}
\usage{
cheb.val(x.axis, cheb.coeff)
}
\description{
cheb.val berechnet aus den Chebyshev-Koeffizienten die Y-Werte
}
\examples{
cheb.val(x.axis, cheb.coeff)
}

