% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-chebyshev.r
\name{cheb.deriv.2nd}
\alias{cheb.deriv.2nd}
\title{Calculation of the values of the second derivation}
\usage{
cheb.deriv.2nd(x.axis, cheb.coeff)
}
\arguments{
\item{x.axis}{beliebige X-Achse (Skalar oder Vektor)}

\item{cheb.coeff}{Chebyshev-Koeffizienten aus Least-Squares-Verfahren (Vektor)}
}
\value{
cheb.model.deriv Zweite Ableitung des gefilterten Modells (Skalar oder Vektor)
}
\description{
cheb.deriv.2nd berechnet aus den Chebyshev-Koeffizienten die Werte der zweiten Ableitung
}
\examples{
cheb.model.deriv.2nd <- cheb.deriv.2nd(x.axis, cheb.coeff)
}

