% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-chebyshev.r
\name{cheb.scale}
\alias{cheb.scale}
\title{Scaling of X-Axis}
\usage{
cheb.scale(x.axis)
}
\arguments{
\item{x.axis}{ursprüngliche beliebige X-Achse (Vektor)}
}
\value{
x.cheb.scaled skalierte X-Achse (Vektor)
}
\description{
\code{cheb.scale} skaliert beliebige X-Achse auf Achse, die für Polynom-fits verträglich ist.
}
\examples{
x.axis <- c(0:30)
x.cheb.scaled <- cheb.scale(x.axis)
}

